39 - Fixed the Maven build problems.  My pom.xml file was missing the <dependencies> section (and had only the <dependencyManagement> section).
I moved RootLayout.fxml to resources/view/ so Maven would copy it to the correct target/ location.  This required an update to Main.java FXMLLoader.load() line.  IntelliJ warns about this line now, saying the Argument "might be null".  I have yet to resolve this.
The old line was:
	Parent root = FXMLLoader.load( getClass().getResource( "view/RootLayout.fxml" ) );
The new line is:
	Parent root = FXMLLoader.load( getClass().getClassLoader().getResource( "view/RootLayout.fxml" ) );